// mobile/models/Color.ts
export enum Color {
    Black = "#1A1A1A",
    White = "white",
    Blue = "#3b82f6",
    Green = "#1DB954",
    Grey = "#B3B3B3",
    DarkGray = "#3A3B3C",
    Red = "#FF6B6B",
    ProfileBlue = "#3B82F6",
    ProfileRed = "#EF4444",
    ProfileYellow = "#F59E0B",
    ProfileGreen = "#10B981",
    ProfileGray = "#2D2D2D"
}